Adding some more unit test for existing functionality (task #5394)
