+++ src/Model/Table/CalendarAttendeesTable.php
@@ -51,7 +51,6 @@ class CalendarAttendeesTable extends Table
 
         $this->addBehavior('Timestamp');
         $this->addBehavior('Muffin/Trash.Trash');
-        $this->addBehavior('AuditStash.AuditLog');
 
         $this->belongsToMany('CalendarEvents', [
             'joinTable' => 'events_attendees',
