+++ src/Model/Table/CalendarsTable.php
@@ -15,13 +15,12 @@ use Cake\Core\Configure;
 use Cake\Datasource\EntityInterface;
 use Cake\Event\Event;
 use Cake\Event\EventManager;
-use Cake\ORM\Query;
 use Cake\ORM\RulesChecker;
 use Cake\ORM\Table;
 use Cake\ORM\TableRegistry;
+use Cake\Utility\Hash;
 use Cake\Validation\Validator;
 use Qobo\Calendar\Event\EventName;
-use Qobo\Calendar\Model\Table\CalendarEventsTable;
 use \ArrayObject;
 
 /**
@@ -59,7 +58,6 @@ class CalendarsTable extends Table
 
         $this->addBehavior('Timestamp');
         $this->addBehavior('Muffin/Trash.Trash');
-        $this->addBehavior('AuditStash.AuditLog');
 
         $this->hasMany('CalendarEvents', [
             'foreignKey' => 'calendar_id',
@@ -123,7 +121,7 @@ class CalendarsTable extends Table
 
         // Default calendar color in case none is given.
         if (empty($entity->color)) {
-            $entity->color = Configure::read('Calendar.Configs.color');
+            $entity->color = $this->getColor($entity);
         }
     }
 
@@ -182,13 +180,30 @@ class CalendarsTable extends Table
     }
 
     /**
+     * Get Default calendar color.
+     *
+     * @param \Cake\ORM\Entity $entity of the current calendar
+     * @return string $color containing hexadecimal color notation.
+     */
+    public function getColor($entity = null)
+    {
+        $color = Configure::read('Calendar.Configs.color');
+
+        if (!$color) {
+            $color = '#337ab7';
+        }
+
+        return $color;
+    }
+
+    /**
      * Get Calendar Types
      *
      * @param array $options with extra filters
      *
      * @return array $result containing calendar types.
      */
-    public function getCalendarTypes($options = [])
+    public function getTypes($options = [])
     {
         $result = [];
 
@@ -198,9 +213,7 @@ class CalendarsTable extends Table
             return $result;
         }
 
-        foreach ($config as $k => $val) {
-            $result[$val['value']] = $val['name'];
-        }
+        $result = Hash::combine($config, '{n}.value', '{n}.name');
 
         return $result;
     }
@@ -212,7 +225,7 @@ class CalendarsTable extends Table
      *
      * @return array $result of the synchronize method.
      */
-    public function syncCalendars($options = [])
+    public function sync(array $options = [])
     {
         $result = [];
 
@@ -234,404 +247,7 @@ class CalendarsTable extends Table
             if (empty($calendar)) {
                 continue;
             }
-
-            // we don't pass period as it doesn't have time limits.
-            $diffCalendar = $this->getItemDifferences(
-                $this,
-                $calendar
-            );
-
-            $result['modified'][] = $this->saveItemDifferences($this, $diffCalendar);
         }
-
-        return $result;
-    }
-
-    /**
-     * Synchronize calendar events
-     *
-     * @param \Cake\ORM\Entity $calendar instance from the db
-     * @param array $data with extra configs
-     *
-     * @return array $result with events responses.
-     */
-    public function syncEventsAttendees($calendar, $data = [])
-    {
-        $result = [];
-        $table = TableRegistry::get('Qobo/Calendar.CalendarEvents');
-        $attendeeTable = TableRegistry::get('Qobo/Calendar.CalendarAttendees');
-
-        if (empty($data)) {
-            return $result;
-        }
-
-        foreach ($data['modified'] as $k => $item) {
-            if (empty($item['attendees'])) {
-                continue;
-            }
-
-            foreach ($item['attendees'] as $attendee) {
-                $diff = $this->getAttendeeDifferences(
-                    $attendeeTable,
-                    $attendee,
-                    [
-                        'source_id' => 'contact_details',
-                    ]
-                );
-                $savedAttendee = $this->saveAttendeeDifferences($attendeeTable, $diff, [
-                    'entity_options' => [
-                        'associated' => ['CalendarEvents'],
-                    ],
-                    'extra_fields' => [
-                        'calendar_events' => [
-                            [
-                                'id' => $item->id,
-                                '_joinData' => [
-                                    'response_status' => $attendee['response_status'],
-                                ]
-                            ]
-                        ],
-                    ],
-                ]);
-
-                $result['modified'][] = $savedAttendee;
-            }
-        }
-
-        return $result;
-    }
-
-    /**
-     * Synchronize calendar events
-     *
-     * @param \Cake\ORM\Entity $calendar instance from the db
-     * @param array $options with extra configs
-     *
-     * @return array $result with events responses.
-     */
-    public function syncCalendarEvents($calendar, $options = [])
-    {
-        $result = [];
-        $table = TableRegistry::get('Qobo/Calendar.CalendarEvents');
-
-        if (empty($calendar)) {
-            return $result;
-        }
-
-        $event = new Event((string)EventName::PLUGIN_CALENDAR_MODEL_GET_EVENTS(), $this, [
-            'calendar' => $calendar,
-            'options' => $options,
-        ]);
-
-        EventManager::instance()->dispatch($event);
-
-        $calendarEvents = $event->result;
-        if (empty($calendarEvents)) {
-            return $result;
-        }
-
-        foreach ($calendarEvents as $k => $calendarInfo) {
-            if (empty($calendarInfo['events'])) {
-                continue;
-            }
-
-            foreach ($calendarInfo['events'] as $item) {
-                $diff = $this->getItemDifferences(
-                    $table,
-                    $item,
-                    $options
-                );
-
-                $savedDiff = $this->saveItemDifferences($table, $diff, [
-                    'extra_fields' => [
-                        'calendar_id' => $calendarInfo['calendar']->id
-                    ],
-                ]);
-
-                $result['modified'][] = $savedDiff;
-            }
-        }
-
-        return $result;
-    }
-
-    /**
-     * saveCalendarDifferences method
-     *
-     * Updating calendar DB with differences
-     *
-     * @param \Cake\ORM\Table $table of the instance
-     * @param array $diff prepopulated calendars
-     * @param array $options with extra configs if any.
-     *
-     * @return array $result with updated/added calendars.
-     */
-    public function saveItemDifferences($table, $diff = [], $options = [])
-    {
-        $result = [];
-        $entityOptions = [];
-
-        if (empty($diff)) {
-            return $result;
-        }
-
-        foreach ($diff as $actionName => $items) {
-            if (empty($items)) {
-                continue;
-            }
-
-            foreach ($items as $k => $item) {
-                $data = [];
-
-                if (empty($item)) {
-                    continue;
-                }
-
-                switch ($actionName) {
-                    case 'add':
-                        $entity = $table->newEntity();
-                        $data = $item;
-                        break;
-                    case 'update':
-                        $entity = $item['entity'];
-                        $data = $item['data'];
-                        break;
-                }
-
-                if (in_array($actionName, ['add', 'update']) && !empty($data)) {
-                    if (!empty($options['extra_fields'])) {
-                        $data = array_merge($data, $options['extra_fields']);
-                    }
-
-                    if (!empty($options['entity_options'])) {
-                        $entityOptions = array_merge($entityOptions, $options['entity_options']);
-                    }
-
-                    $entity = $table->patchEntity($entity, $data, $entityOptions);
-                    $result = $table->save($entity);
-                }
-            }
-        }
-
-        return $result;
-    }
-
-    /**
-     * Save Attendee Differences
-     *
-     * Checkes whether attendee should be added/updated/removed
-     *
-     * @param \Cake\ORM\Table $table instance
-     * @param array $diff containing the data
-     * @param array $options with extra settings/fields to save/modify
-     *
-     * @return array $result containing diff results
-     */
-    public function saveAttendeeDifferences($table, $diff = [], $options = [])
-    {
-        $result = [];
-        $entityOptions = [];
-
-        if (empty($diff)) {
-            return $result;
-        }
-
-        foreach ($diff as $actionName => $items) {
-            if (empty($items)) {
-                continue;
-            }
-
-            foreach ($items as $k => $item) {
-                $data = [];
-
-                if (empty($item)) {
-                    continue;
-                }
-
-                switch ($actionName) {
-                    case 'add':
-                        $entity = $table->newEntity();
-                        $data = $item;
-                        break;
-                    case 'update':
-                        $entity = $item['entity'];
-                        $data = $item['data'];
-                        break;
-                }
-
-                if (in_array($actionName, ['add', 'update']) && !empty($data)) {
-                    if (!empty($options['extra_fields'])) {
-                        $data = array_merge($data, $options['extra_fields']);
-                    }
-
-                    if (!empty($options['entity_options'])) {
-                        $entityOptions = array_merge($entityOptions, $options['entity_options']);
-                    }
-
-                    $entity = $table->patchEntity($entity, $data, $entityOptions);
-                    $savedAttendee = $table->save($entity);
-                }
-            }
-        }
-
-        return $result;
-    }
-
-    /**
-     * Collect calendars difference.
-     *
-     * @param \Cake\ORM\Table $table related instance.
-     * @param array $item to be checked for add/update (aka calendar or event).
-     * @param array $options with extra configs
-     *
-     * @return $calendarDiff containing the list of calendars to add/update.
-     */
-    public function getItemDifferences($table, $item = null, $options = [])
-    {
-        $conditions = [];
-        $source = empty($options['source']) ? 'source' : $options['source'];
-        $sourceId = empty($options['source_id']) ? 'source_id' : $options['source_id'];
-
-        $diff = [
-            'add' => [],
-            'update' => [],
-        ];
-
-        if (empty($item)) {
-            return $diff;
-        }
-
-        if (is_null($item[$source])) {
-            $conditions[$source . ' IS'] = $item[$source];
-        } else {
-            $conditions[$source] = $item[$source];
-        }
-
-        $conditions[$sourceId] = $item[$sourceId];
-
-        $query = $table->find()
-                ->where($conditions);
-
-        $query->all();
-        $dbItems = $query->toArray();
-
-        $toAdd = $this->itemsToAdd($item, $dbItems, $sourceId);
-        if (!empty($toAdd)) {
-            $diff['add'][] = $toAdd;
-        }
-
-        $toUpdate = $this->itemsToUpdate($item, $dbItems, $sourceId);
-        if (!empty($toUpdate)) {
-            $diff['update'][] = $toUpdate;
-        }
-
-        return $diff;
-    }
-
-    /**
-     * Get Attendee difference
-     *
-     * @param \Cake\ORM\Table $table instance of attendees
-     * @param array $item of the record
-     * @param array $options for extra fields/conditions
-     *
-     * @return array $diff with sorted differences for the item.
-     */
-    public function getAttendeeDifferences($table, $item = null, $options = [])
-    {
-        $conditions = [];
-        $sourceId = empty($options['source_id']) ? 'source_id' : $options['source_id'];
-
-        $diff = [
-            'add' => [],
-            'update' => [],
-        ];
-
-        if (empty($item)) {
-            return $diff;
-        }
-
-        $conditions[$sourceId] = $item[$sourceId];
-
-        $query = $table->find()
-                ->where($conditions);
-
-        $query->all();
-        $dbItems = $query->toArray();
-
-        $toAdd = $this->itemsToAdd($item, $dbItems, $sourceId);
-        if (!empty($toAdd)) {
-            $diff['add'][] = $toAdd;
-        }
-
-        $toUpdate = $this->itemsToUpdate($item, $dbItems, $sourceId);
-        if (!empty($toUpdate)) {
-            $diff['update'][] = $toUpdate;
-        }
-
-        return $diff;
-    }
-
-    /**
-     * Check if calendar should be added
-     *
-     * @param array $item to inspect for adding
-     * @param array $dbItems to be compared with
-     * @param string $fieldToCheck lookup field name
-     *
-     * @return array $result with the comparison result.
-     */
-    public function itemsToAdd($item, $dbItems = [], $fieldToCheck = null)
-    {
-        $result = $item;
-
-        if (empty($dbItems)) {
-            return $result;
-        }
-
-        foreach ($dbItems as $k => $dbItem) {
-            if ($dbItem->$fieldToCheck == $item[$fieldToCheck]) {
-                $result = [];
-                break;
-            }
-        }
-
-        return $result;
-    }
-
-    /**
-     * Check if the calendar should be updated
-     *
-     * @param array $item to be checked for update
-     * @param array $dbItems to be checked against
-     * @param string $fieldToCheck lookup field name
-     *
-     * @return array $result containing comparison result
-     */
-    public function itemsToUpdate($item, $dbItems = [], $fieldToCheck = null)
-    {
-        $found = null;
-        $result = [
-            'entity' => [],
-            'data' => []
-        ];
-
-        if (empty($dbItems)) {
-            return $result;
-        }
-
-        foreach ($dbItems as $dbItem) {
-            if ($dbItem->$fieldToCheck == $item[$fieldToCheck]) {
-                $found = $dbItem;
-            }
-        }
-
-        if (empty($found)) {
-            return $result;
-        }
-
-        $result['entity'] = $found;
-        $result['data'] = $item;
 
         return $result;
     }
